refactor: extract earth model logic to ocean page
This make it more flexible to reuse the model without
that logic, which is specific of the ocean page.
