--- oceangarden/src/app/model/page.tsx
@@ -1,9 +1,0 @@
-"use client"
-
-import Model from "@/components/Model3D/Model"
-
-const Home = () => {
-  return <Model />
-}
-
-export default Home
